+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Apollo Link State support.
+- 
 ## [8.0.6] - 2019-1-22
 ### Changed
 - Loading svg is loaded by Icon Pack.